requirement.txt
altgraph==0.17.4
chardet==5.2.0
importlib_metadata==7.2.0
packaging==24.1
pefile==2023.2.7
pillow==10.3.0
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pywin32-ctypes==0.2.2
reportlab==4.2.0
zipp==3.19.2
